% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chol_inv}
\alias{chol_inv}
\title{Calculates the inverse of the Cholesky factor via the
regression coefficients formula}
\usage{
chol_inv(coeffs)
}
\arguments{
\item{coeffs}{Lower-triangular matrix of regression coefficients}
}
\value{
Matrix inverse of the Cholesky factor
}
\description{
Calculates the inverse of the Cholesky factor via the
regression coefficients formula
}
