% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cholfromldl}
\alias{cholfromldl}
\title{Returns Cholesky factorization from ldl}
\usage{
cholfromldl(L, D)
}
\arguments{
\item{L}{L factor in LDL^t decomposition}

\item{D}{D factor (diagonal vector) in LDL^t decomposition}
}
\value{
The lower Cholesky factor
}
\description{
Returns Cholesky factorization from ldl
}
