% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_chol_conc}
\alias{fit_chol_conc}
\title{Fit the Choleksy factor of the inverse covariance / concentration matrix 
with linear regression}
\usage{
fit_chol_conc(dag, data)
}
\arguments{
\item{dag}{Graph marking the sparsity of the factor}

\item{data}{Data for fitting}
}
\value{
Cholesky factor  of the inverse covariance matrix
}
\description{
Fit the Choleksy factor of the inverse covariance / concentration matrix 
with linear regression
}
