% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_chol_cov}
\alias{fit_chol_cov}
\title{Fit the Cholesky factor of the covariance matrix with linear regression}
\usage{
fit_chol_cov(amat, data)
}
\arguments{
\item{amat}{Adjacency matrix marking the sparsity of the factor}

\item{data}{Data for fitting}
}
\value{
Cholesky factor of the covariance matrix
}
\description{
Fit the Cholesky factor of the covariance matrix with linear regression
}
