% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_ldl_cov}
\alias{fit_ldl_cov}
\title{Fit the ldl factor of the covariance matrix with linear regression}
\usage{
fit_ldl_cov(amat, data)
}
\arguments{
\item{amat}{Adjacency matrix marking the sparsity of the factor}

\item{data}{Data for fitting}
}
\value{
A list with the ldl factors
}
\description{
Fit the ldl factor of the covariance matrix with linear regression
}
