% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ldlfromchol}
\alias{ldlfromchol}
\title{Returns LDL factorization from lower Cholesky factor}
\usage{
ldlfromchol(L)
}
\arguments{
\item{L}{Lower Cholesky factor}
}
\value{
a list with the L and D factors (D being a vector)
}
\description{
Returns LDL factorization from lower Cholesky factor
}
