% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxgrad.R
\name{llpathL}
\alias{llpathL}
\title{Path of L estimates (log-lik)}
\usage{
llpathL(
  Sigma,
  lambdas = NULL,
  D = diag(nrow(Sigma)),
  L0 = NULL,
  eps = 1e-08,
  maxIter = 1000,
  r = FALSE,
  h = TRUE
)
}
\arguments{
\item{Sigma}{empirical covariance matrix}

\item{lambdas}{increasing sequence of lambdas}

\item{D}{the known noise matrix}

\item{L0}{initial L matrix}

\item{eps}{convergence criteria}

\item{maxIter}{maximum iterations for each proximal gradient}

\item{r}{as in proxgradL}

\item{h}{as in proxgradL}
}
\description{
Path of L estimates (log-lik)
}
