% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxgrad.R
\name{proxgradL}
\alias{proxgradL}
\title{Proximal gradient sparse Cholesky factor}
\usage{
proxgradL(
  Sigma,
  L,
  D = diag(ncol(Sigma)),
  eps = 1e-04,
  alpha = 0.2,
  beta = 0.5,
  maxIter = 100,
  trace = 0,
  lambda = 0,
  r = FALSE,
  h = TRUE
)
}
\arguments{
\item{Sigma}{the empirical covariance matrix}

\item{L}{an initial guess for L}

\item{D}{the noise matrix}

\item{eps}{stopping criteria}

\item{alpha}{parameter for line search}

\item{beta}{parameter for line search}

\item{maxIter}{maximum number of iterations}

\item{trace}{if >0 print info (1:termination message only, 
2:messages at each step)}

\item{lambda}{penalization coefficient}

\item{r}{logical, if TRUE the set of non-zero parameters will be updated 
at every iteration}

\item{h}{logical if TRUE only the non-zero entries of L will be updated}
}
\value{
A list with
        * L 
        * Sigma = LDL^t
        * niter 
        * diff
}
\description{
Proximal gradient sparse Cholesky factor
}
