% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prxgradl.R
\name{prxgradl}
\alias{prxgradl}
\title{Penalized likelihood estimation of Cholesky factor}
\usage{
prxgradl(Sigma, L, eps = 0.01, alpha = 0.5, maxIter = 100, lambda = 0)
}
\arguments{
\item{Sigma}{the observed covariance matrix}

\item{L}{an initial L matrix}

\item{eps}{convergence threshold for the proximal gradient}

\item{alpha}{Beck and Tabulle line search rate}

\item{maxIter}{the maximum number of iterations}

\item{lambda}{penalization coefficient}
}
\value{
a list with the output of the optimization:

* \code{N}
* \code{L} the estimated L matrix
* \code{lambda} 
* \code{diff} the value of the last relative decrease
* \code{objective} the value of the objective function
* \code{iter} number of iterations
}
\description{
Optimize the  matrix of a continuous Lyapunov 
Gaussian graphical model (CLGGM) using proximal gradient. 
\deqn{\hat{L} = \arg \min_L \log + \lambda||L||_1}
}
