% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlower}
\alias{rlower}
\title{Simulates a lower triangular matrix, possibly with a sparsity pattern}
\usage{
rlower(p, d)
}
\arguments{
\item{p}{Dimension of the matrix to be simulated}

\item{d}{Density degree}
}
\value{
The simulated lower triangular matrix
}
\description{
Simulates a lower triangular matrix, possibly with a sparsity pattern
}
\details{
Each entry in the lower triangle of the matrix is first assigned
the result of a Bernouilli with parameter `d`. If a `1` was obtained, then
the entry is assigned the result of a Uniform distribution on `[0.1, 1]`.
}
